update for gm skill gift and 0.00 skill level fix

+++ src/eve-server/character/CharUnboundMgrService.cpp
@@ -315,7 +315,7 @@ PyResult CharUnboundMgrService::Handle_CreateCharacterWithDoll(PyCallArgs &call)
         }
 
         skillLevel = cur.second;
-        skill->SetAttribute(AttrSkillLevel, (int8)skillLevel, false);
+        skill->SetAttribute(AttrSkillLevel, skillLevel, false);
         skillPoints = skill->GetSPForLevel( (EvilNumber)skillLevel );
         skill->SetAttribute(AttrSkillPoints, skillPoints, false);
         skill->SaveItem();
